+++ src/transport/dbus.rs
@@ -1,0 +1,3 @@
+//! D-Bus support. Currently unimplemented.
+//!
+//! D-Bus support will be implemented here, likely according to [this description](https://github.com/victronenergy/venus/wiki/Venus-OS-D%E2%80%90Bus-S2-Interface). It is not currently available.